Move tests to a more appropriate place
